Refactoring GATT
